Refactoring folder names
