+++ src/main/java/com/example/spring_boot/Application.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-@EnableFeignClients(basePackages = "com.example.spring_boot.client")
+@EnableFeignClients(basePackages = "com.example.spring_boot.clients")
 @EnableScheduling // 🔥 Habilita las tareas programadas
 public class Application {
