+++ src/main/java/com/example/spring_boot/services/GhRepositoryManager.java
@@ -3,9 +3,8 @@ package com.example.spring_boot.services;
 import com.example.spring_boot.dtos.GhRepositoryRequest;
 import com.example.spring_boot.models.GhRepository;
 import com.example.spring_boot.repositories.GhRepositoryRepository;
-import com.example.spring_boot.client.GhClient;
+import com.example.spring_boot.clients.GhClient;
 import feign.FeignException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
